backend/requirements.txt
# Core Deep Learning
tensorflow==2.20.0

# Data Science
numpy==1.26.3
pandas==2.2.0

# Image Processing  
Pillow==10.2.0
opencv-python==4.9.0.80

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Scientific Computing
scikit-learn==1.4.0